+++ NEWS.md
@@ -1,5 +1,8 @@
 # mizer 2.0.4.9000
 
+* New function `projectToSteady()` to run the full dynamics to steady state.
+* New functions `distanceSSLogN()` and `distanceMaxRelRDI()` to measure 
+  distance between two states.
 * Improved error handling in `setFishing()` #172
 * Now if the effort is specified as a named vector giving values only for some 
   gears, the effort for the remaining gears is assumed to be zero.